Patch edge case
I do not expect this to have occured much:
This issue requires the chain to be empty,
and requires there to be elements that are not in use.
This means this will not occur in the first iteration
as all elements are in use.

Emptying out the chain requires finding a reciprocal point in the first 3 items.
In this case a random element should be selected that is no longer in use.

Finally, I added a check to make sure the input matrix is symmetric.
An asymmetric matrix causes some issues, most notably with
checking if the chain is reciprocal.

This causes the chain to fill up until it hits the extra case:
`length(NN_chain) == n`
where the items are automatically assumed to be reciprocal.
The chain is however likely invalid as this point:
many indices occur more than once, not just the reciprocal one.