Set up experiments for re-running GOMEAs
Due to an error the versions of GOMEA implemented in Julia
have been using a steady-state like variant. This is not how
the original Permutation GOMEA - nor qGOMEA was supposed
to work. A previous commit fixed this issue.
This commit contains the required scripts to re-run experiments
which were affected by this fix (all experiments involving GOMEA).
Furthermore, as this re-run will be completely performed on a server,
all runs for BIV (the inversions benchmark functions) will be re-ran again,
as otherwise hardware differences make the wall-time limit questionable.