Disable ER again
As per the partial run (that was aborted by a machine
failure?) the performance of this approach is still too
subpar. As such, we still do not include this in our tables.
Given the deadline, no need to run an approach that is
not going to be included in the article.